move ProjectData to own File
