+++ catroid/src/org/catrobat/catroid/ui/adapter/ProjectAdapter.java
@@ -36,8 +36,8 @@ import android.widget.ListView;
 import android.widget.TextView;
 
 import org.catrobat.catroid.R;
+import org.catrobat.catroid.common.ProjectData;
 import org.catrobat.catroid.io.ProjectScreenshotLoader;
-import org.catrobat.catroid.ui.fragment.ProjectsListFragment.ProjectData;
 import org.catrobat.catroid.utils.UtilFile;
 import org.catrobat.catroid.utils.Utils;
